Entities Now Have Double-Sized Positions

+++ Main/Entity.cpp
@@ -2,18 +2,18 @@
 using namespace std;
 
 //Simple Constructor
-Entity::Entity(const unsigned int x, const unsigned int y, Bin* const bin)
+Entity::Entity(const double x, const double y, Bin* const bin)
   : x(x), y(y), bin(bin) { }
 
 //Deconstructor for derived classes
 Entity::~Entity() { }
 
-unsigned int Entity::getX() const
+double Entity::getX() const
 {
     return x;
 }
 
-unsigned int Entity::getY() const
+double Entity::getY() const
 {
     return y;
 }
